#51 [modify] 스크롤 관련 버그 수정
