--- src/allTypes/myStudy.ts
@@ -1,6 +1,0 @@
-export interface IDiary {
-  diaryId: number;
-  userId: number;
-  content: string;
-  createdAt: string;
-}
